# core_tron_cmn_api/1-DEVELOPMENT/javaApp/nwt_cmn_api_be-sr-api/src/main/java/com/mapfre/tron/api/cmn/sr/ThirdLevelDistributionChannelByCmpFirstSecondApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.14).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.mapfre.tron.api.cmn.sr;

import java.io.IOException;
import java.util.List;
import java.util.Optional;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import javax.validation.constraints.NotNull;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.mapfre.nwt.trn.dsr.hdc.bo.ODsrHdcS;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import io.swagger.annotations.Authorization;

@Api(value = "thirdLevelDistributionChannelByCmpFirstSecond", description = "the thirdLevelDistributionChannelByCmpFirstSecond API")
@RequestMapping(value = "/newtron/api/common")
public interface ThirdLevelDistributionChannelByCmpFirstSecondApi {

    Logger log = LoggerFactory.getLogger(ThirdLevelDistributionChannelByCmpFirstSecondApi.class);

    default Optional<ObjectMapper> getObjectMapper() {
        return Optional.empty();
    }

    default Optional<HttpServletRequest> getRequest() {
        return Optional.empty();
    }

    default Optional<String> getAcceptHeader() {
        return getRequest().map(r -> r.getHeader("Accept"));
    }

    @ApiOperation(value = "Third Level Distribution Channel by company, first level and second level", nickname = "getthirdLevelDistributionChannelByCmpFirstSecond", notes = "Query Third Level Distribution Channel by company, first level and second level. It will return the types of languages. It will be mandatory send the parameters defined in the service and the service will response with the output object defined.", response = ODsrHdcS.class, responseContainer = "List", authorizations = {
        @Authorization(value = "basicAuth")
    }, tags={ "Common", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Language List", response = ODsrHdcS.class, responseContainer = "List"),
        @ApiResponse(code = 400, message = "Bad request", response = Error.class),
        @ApiResponse(code = 404, message = "Not found", response = Error.class),
        @ApiResponse(code = 500, message = "Internal error", response = Error.class) })
    @RequestMapping(value = "/thirdLevelDistributionChannelByCmpFirstSecond",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<ODsrHdcS>> getthirdLevelDistributionChannelByCmpFirstSecond(@NotNull @ApiParam(value = "Company code", required = true) @Valid @RequestParam(value = "cmpVal", required = true) Integer cmpVal,@NotNull @ApiParam(value = "First Channel Distribution", required = true) @Valid @RequestParam(value = "frsDstHnlVal", required = true) String frsDstHnlVal,@NotNull @ApiParam(value = "Second Channel Distribution", required = true) @Valid @RequestParam(value = "scnDstHnlVal", required = true) String scnDstHnlVal,@ApiParam(value = "Language code" ,required=true) @RequestHeader(value="lngVal", required=true) String lngVal,@NotNull @ApiParam(value = "User code", required = true) @Valid @RequestParam(value = "usrVal", required = true) String usrVal) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("[ {  \"frsDstHnlNam\" : \"frsDstHnlNam\",  \"dsbRow\" : \"dsbRow\",  \"thrDstHnlAbr\" : \"thrDstHnlAbr\",  \"thrDstHnlVal\" : \"thrDstHnlVal\",  \"scnDstHnlVal\" : \"scnDstHnlVal\",  \"cmpVal\" : 0.80082819046101150206595775671303272247314453125,  \"scnDstHnlNam\" : \"scnDstHnlNam\",  \"frsDstHnlVal\" : \"frsDstHnlVal\",  \"thrDstHnlNam\" : \"thrDstHnlNam\"}, {  \"frsDstHnlNam\" : \"frsDstHnlNam\",  \"dsbRow\" : \"dsbRow\",  \"thrDstHnlAbr\" : \"thrDstHnlAbr\",  \"thrDstHnlVal\" : \"thrDstHnlVal\",  \"scnDstHnlVal\" : \"scnDstHnlVal\",  \"cmpVal\" : 0.80082819046101150206595775671303272247314453125,  \"scnDstHnlNam\" : \"scnDstHnlNam\",  \"frsDstHnlVal\" : \"frsDstHnlVal\",  \"thrDstHnlNam\" : \"thrDstHnlNam\"} ]", List.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default ThirdLevelDistributionChannelByCmpFirstSecondApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}