# core_tron_cmn_api/1-DEVELOPMENT/javaApp/nwt_cmn_api_be-sr-impl/src/main/java/com/mapfre/tron/api/cmn/sr/ListsController.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.14).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.mapfre.tron.api.cmn.sr;

import java.util.List;

import javax.validation.Valid;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestParam;

import com.mapfre.nwt.trn.trn.lst.bo.OTrnLstValS;
import com.mapfre.tron.api.cmn.lst.bl.IBlCmnLstQry;
import com.mapfre.tron.api.cmn.model.InInfoList;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiParam;
import lombok.Data;
import lombok.extern.slf4j.Slf4j;

/**
 * The Values Option lists controller.
 *
 * @author parlaga
 * @since 1.8
 * @version 20 oct. 2020 9:30:02
 *
 */
@Data
@Slf4j
@Controller
@Api(tags={ "Values/options list",})
public class ListsController implements ListsApi {

    @Autowired
    private IBlCmnLstQry blCmnLstQry;
    
    @Override
    public ResponseEntity<List<OTrnLstValS>> getValuesOptionList(@ApiParam(value = "Input data to get lists" ,required=true )  @Valid @RequestBody InInfoList inInfoList, @ApiParam(value = "Company code" ,required=false) @RequestParam(value="cmpVal", required=false) Integer cmpVal) {
       log.info("The operation return option/values list is called...");
       
	   List<OTrnLstValS> optionValuesList = blCmnLstQry.getLstVal(inInfoList.getLstIdn(),inInfoList.getLstTyp(), inInfoList.getInsVal(), inInfoList.getLstVrs(), inInfoList.getLstVal(), inInfoList.isCchChc(), cmpVal);
		       
           return ResponseEntity.ok(optionValuesList);
    }

}