# core_tron_cmn_api/1-DEVELOPMENT/swagger/CMN_API_NEWTRON.yaml
swagger: '2.0'
info:
  description: Mapfre CMN API NEWTRON
  version: '${project.version}'
  title: Mapfre CMN API NEWTRON
basePath: /newtron/api/common
host: api.mapfre.com
securityDefinitions:
  basicAuth:
    type: basic
security:
  - basicAuth: []
schemes:
  - http
  - https
paths:
...
  /lists:
    post:
      tags:
        - Values/options list
      summary: Get tron lists 
      description: >-
        Service that allows you to retrieve the information of all
        lists of values ​​or options.
      operationId: getValuesOptionList
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: inInfoList
          in: body
          description: Input data to get lists
          required: true
          schema:
            $ref: '#/definitions/InInfoList'
        - name: cmpVal
          in: query
          description: Company code
          required: false
          type: integer 
      responses:
        '201':
          description: Get Lists
          schema:
            type: array            
            items:
              $ref: 'API_OBJECTS.yaml#/definitions/OTrnLstValS'
        '400':
          description: Bad request
          schema:
            $ref: 'API_OBJECTS.yaml#/definitions/Error'
        '404':
          description: Not found
          schema:
            $ref: 'API_OBJECTS.yaml#/definitions/Error'
        '500':
          description: Internal error
          schema:
            $ref: 'API_OBJECTS.yaml#/definitions/Error'
...
  /thirdLevelDistributionChannelByCmpFirstSecond:
    get:
      tags:
        - Common
      summary: Third Level Distribution Channel by company, first level and second level
      description: >-
        Query Third Level Distribution Channel by company, first level and second level.
        It will return the types of languages. It will be
        mandatory send the parameters defined in the service and the service
        will response with the output object defined.
      operationId: getthirdLevelDistributionChannelByCmpFirstSecond
      produces:
        - application/json
      parameters:
        - name: cmpVal
          in: query
          description: Company code
          required: true
          type: integer
        - name: frsDstHnlVal
          in: query
          description: First Channel Distribution
          required: true
          type: string       
        - name: scnDstHnlVal
          in: query
          description: Second Channel Distribution
          required: true
          type: string               
        - name: lngVal
          in: header
          description: Language code
          required: true
          type: string
        - name: usrVal
          in: query
          description: User code
          required: true
          type: string
      responses:
        '200':
          description: Language List
          schema:
            type: array
            items:
              $ref: 'API_OBJECTS.yaml#/definitions/ODsrHdcS'
        '400':
          description: Bad request
          schema:
            $ref: 'API_OBJECTS.yaml#/definitions/Error'
        '404':
          description: Not found
          schema:
            $ref: 'API_OBJECTS.yaml#/definitions/Error'
        '500':
          description: Internal error
          schema:
            $ref: 'API_OBJECTS.yaml#/definitions/Error'
...
definitions:
...
  InInfoList:
    description: Input Option Values list.
    properties:
      lstIdn:
        type: string
        description: List Identifier
        example: 'Sec'
      lstTyp:
        type: string
        description: List type
        example: 'V'
      insVal:
        type: string
        description: Installation value
        example: 'TRN'
      lstVrs:
        type: string
        description: List version
        example: '2'
      lstVal:
        description: Properties list
        type: object
      cchChc:
        type: boolean
        description: Cacheable
        example: true
...